Lazify `hyperspy.misc.signal_tools`
